menambahkan fitur autorisasi

+++ src/api/notes/handler.js
@@ -13,8 +13,14 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = "untitled", body, tags } = request.payload;
+    const { id: credentialId } = request.auth.credentials;
 
-    const noteId = await this._service.addNote({ title, body, tags });
+    const noteId = await this._service.addNote({
+      title,
+      body,
+      tags,
+      owner: credentialId,
+    });
 
     const response = h.response({
       status: "success",
@@ -27,8 +33,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: "success",
       data: {
@@ -39,22 +46,25 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
-    return h
-      .response({
-        status: "success",
-        data: {
-          note,
-        },
-      })
-      .code(200);
+    return {
+      status: "success",
+      data: {
+        note,
+      },
+    };
   }
 
   async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
-    await this._service.editNoteById(id, request.payload);
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.editNoteById(id, request.payload);
 
     return {
       status: "success",
@@ -64,7 +74,10 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
-    await this._service.deleteNoteById(id);
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.deleteNoteById(id);
 
     return {
       status: "success",
